% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritores.R
\name{descritores}
\alias{descritores}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
descritores(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
desc_PAM <- descritores(1612)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

