% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pode_num.R
\name{pode_num}
\alias{pode_num}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
pode_num(x)
}
\arguments{
\item{x}{Um vetor.}
}
\description{
Verifica se um vetor pode ou nao ser convertido em numérico em . Retorna um valor lógico.
}
\examples{
PAM <- API_SIDRA(1612, 1, categoria = 81)
numericos <- sapply(PAM, pode_num)
}
\keyword{coercion}
\keyword{numeric}

