% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRA_classificacao.R
\name{SIDRA_classificacao}
\alias{SIDRA_classificacao}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
SIDRA_classificacao(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
cat_PAM <- SIDRA_classificacao(1612)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

