% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRA_fonte.R
\name{SIDRA_fonte}
\alias{SIDRA_fonte}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
SIDRA_fonte(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
fonte_PAM <- SIDRA_fonte(1612)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

