% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRA_nivel.R
\name{SIDRA_nivel}
\alias{SIDRA_nivel}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
SIDRA_nivel(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
nivel_PAM <- SIDRA_nivel(1612)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

