% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRA_notas.R
\name{SIDRA_notas}
\alias{SIDRA_notas}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
SIDRA_notas(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
nota_PAM <- SIDRA_notas(1612)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

