% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRA_periodo.R
\name{SIDRA_periodo}
\alias{SIDRA_periodo}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
SIDRA_periodo(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
periodo_PAM <- SIDRA_periodo(1612)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

