% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRA_tabelas.R
\name{SIDRA_tabelas}
\alias{SIDRA_tabelas}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
SIDRA_tabelas(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
tabelas <- SIDRA_tabelas()
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

