% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDRA_variaveis.R
\name{SIDRA_variaveis}
\alias{SIDRA_variaveis}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
SIDRA_variaveis(tabela)
}
\arguments{
\item{tabela}{Número da tabela.}
}
\description{
Esta função retorna uma lista com as variáveis e categorias de uma das tabelas da SIDRA.
}
\examples{
var_PAM <- SIDRA_variaveis(1612)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

