% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabelas_SIDRA}
\alias{tabelas_SIDRA}
\title{Tabelas contidas no Sistema IBGE de Recuperação Automática - SIDRA.}
\format{Um data.frame com 5232 linhas e 2 variáveis:
\describe{
  \item{tabela}{Código da tabela}
  \item{descricao}{Nome da tabela}
  ...
}}
\source{
\url{http://api.sidra.ibge.gov.br/}
}
\usage{
tabelas_SIDRA
}
\description{
Um relação do código e descrição das tabelas contidas na SIDRA
}
\keyword{datasets}

