% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_SIDRA.R
\name{API_SIDRA}
\alias{API_SIDRA}
\title{Conexão do R com a SIDRA - IBGE}
\usage{
API_SIDRA(tabela, classificador, cod_cat = "all", nivel = 1,
  cod_nivel = "all", periodo = "all", variavel = "allxp",
  header = FALSE)
}
\arguments{
\item{tabela}{Número da tabela.}

\item{classificador}{Classificador a ser detalhado. Para verificar os classificadores disponíveis na tabela em questão use a função descritores().}

\item{cod_cat}{Código para definição de subconjunto do classificador. Para verificar as categorias disponíveis na tabela em questão use a função descritores().}

\item{nivel}{Nível geográfico de agregação dos dados 1 = Brasil e 6 = Município.}

\item{cod_nivel}{Código contendo conjunto no nível que será selecionado. Pode-se usar o código de determina UF para obter apenas seus dados ou "all" para todos (padrão).}

\item{periodo}{Período dos dados. O padrão é "all", isto é, todos os anos disponíveis.}

\item{variavel}{Quais variáveis devem retornar? O padrão é "allxp", isto é, todas exceto aquelas calculadas pela SIDRA (percentuais)}

\item{header}{Deve retornar com um cabeçalho? Mesmo com header = FALSE, as colunas são nomeadas.}
}
\description{
Esta função retorna um data.frame a tabela solicitada.
}
\examples{
PAM <- API_SIDRA(1612, 81)
}
\keyword{IBGE}
\keyword{SIDRA}
\keyword{dados}

